fix: rename Utils to utils for case-sensitive filesystems
