+++ .clang-tidy
@@ -1,4 +1,3 @@
 Checks:          'modernize-*,performance-*,bugprone-*,readability-*,clang-analyzer-*,cppcoreguidelines-*,misc-*,google-*'
 WarningsAsErrors: ''
 HeaderFilterRegex: 'src/.*'
-FormatStyle:     fileExcludePaths: 'thirdparty/.*'
