+++ src/utils/WindowInfo.cpp
